@@ -65,17 +65,17 @@ extension BSON.ListEncoder
     @inlinable public mutating
     func append(with encode:(inout BSON.ListEncoder) -> ())
     {
-        self.append { $0.encode(with: encode) }
+        self.append { encode(&$0[as: BSON.ListEncoder.self]) }
     }
     @inlinable public mutating
     func append(with encode:(inout BSON.DocumentEncoder<BSON.Key>) -> ())
     {
-        self.append { $0.encode(with: encode) }
+        self.append { encode(&$0[as: BSON.DocumentEncoder<BSON.Key>.self]) }
     }
     @inlinable public mutating
     func append<CodingKeys>(using _:CodingKeys.Type = CodingKeys.self,
         with encode:(inout BSON.DocumentEncoder<CodingKeys>) -> ())
     {
-        self.append { $0.encode(with: encode) }
+        self.append { encode(&$0[as: BSON.DocumentEncoder<CodingKeys>.self]) }
     }
 }
