@@ -1,4 +1,5 @@
 import Durations
+import MongoConfiguration
 import NIOCore
 import NIOPosix
 
@@ -31,7 +32,7 @@ extension Mongo.DriverBootstrap
                 try? await executors.shutdownGracefully()
                 throw error
             }
-        
+
         case .shared(let executors):
             return try await operation(executors)
         }
@@ -72,19 +73,12 @@ extension Mongo.DriverBootstrap
                 connectorFactory: connectorFactory,
                 authenticator: authenticator,
                 deployment: deployment)
-            
-            #if compiler(<5.8)
-            async
-            let __:Void = monitors.start(from: self.seeding,
-                interval: self.monitorInterval,
-                topology: self.topology)
-            #else
+
             async
             let _:Void = monitors.start(from: self.seeding,
                 interval: self.monitorInterval,
                 topology: self.topology)
-            #endif
-            
+
             let sessions:Mongo.SessionPool = .init(deployment: deployment)
             do
             {
