@@ -1,6 +1,6 @@
 import BSONDecoding
 import BSONEncoding
-import Durations
+import MongoDriver
 import NIOCore
 
 extension Mongo
@@ -31,6 +31,21 @@ extension Mongo
         }
     }
 }
+extension Mongo.Find
+{
+    private
+    init(collection:Mongo.Collection,
+        readConcern:ReadConcern?,
+        tailing:Mode.Tailing?,
+        stride:Mode.Stride,
+        with populate:(inout BSON.DocumentEncoder<BSON.Key>) -> ())
+    {
+        self.init(readConcern: readConcern,
+            tailing: tailing,
+            stride: stride,
+            fields: Self.type(collection, then: populate))
+    }
+}
 extension Mongo.Find:MongoImplicitSessionCommand, MongoTransactableCommand, MongoCommand
 {
     /// The string [`"find"`]().
@@ -57,21 +72,6 @@ extension Mongo.Find:MongoIterableCommand
 {
     public
     typealias Element = Mode.Element
-}
-extension Mongo.Find
-{
-    private
-    init(collection:Mongo.Collection,
-        readConcern:ReadConcern?,
-        tailing:Mode.Tailing?,
-        stride:Mode.Stride,
-        with populate:(inout BSON.DocumentEncoder<BSON.Key>) -> ())
-    {
-        self.init(readConcern: readConcern,
-            tailing: tailing,
-            stride: stride,
-            fields: Self.type(collection, then: populate))
-    }
 }
 extension Mongo.Find where Mode.Tailing == Mongo.Tailing, Mode.Stride == Int
 {
