@@ -1,9 +1,10 @@
 import BSONDecoding
+import MongoDriver
 import NIOCore
 
 extension Mongo
 {
-    public
+    @frozen public
     enum SingleBatch<Element> where Element:BSONDecodable & Sendable
     {
     }
